Refactor FSDP Engine and its utilities for future expert parallelism
