+++ areal/engine/fsdp_engine.py
@@ -8,21 +8,13 @@ from typing import Any, Callable, Dict, List
 import torch
 import torch.distributed as dist
 import torch.distributed.nn.functional as dist_F
-import torch.nn as nn
 from torch.distributed.checkpoint.state_dict import (
     StateDictOptions,
     get_model_state_dict,
 )
-from torch.distributed.device_mesh import DeviceMesh, init_device_mesh
-from torch.distributed.tensor import DTensor, Replicate, Shard
-from torch.distributed.tensor.parallel import (
-    ColwiseParallel,
-    ParallelStyle,
-    PrepareModuleInput,
-    RowwiseParallel,
-    SequenceParallel,
-    parallelize_module,
-)
+from torch.distributed.device_mesh import DeviceMesh
+from torch.distributed.fsdp import CPUOffloadPolicy
+from torch.distributed.tensor import DTensor
 from transformers import PreTrainedTokenizerFast, ProcessorMixin
 
 from areal.api.alloc_mode import FSDPParallelStrategy, ParallelStrategy
@@ -39,15 +31,10 @@ from areal.utils.data import (
     unpack_sequence,
 )
 from areal.utils.distributed import init_custom_process_group
-from areal.utils.fsdp import (
-    CPUOffloadPolicy,
-    MixedPrecisionPolicy,
-    NoParallel,
-    apply_fsdp2,
-    fsdp2_clip_grad_norm,
-    fsdp2_load_full_state_dict,
-)
-from areal.utils.model import VALID_VISION_MODELS, is_gemma3_model
+from areal.utils.fsdp import fsdp2_load_full_state_dict
+from areal.utils.fsdp.grad import fsdp2_clip_grad_norm
+from areal.utils.fsdp.parallel_dims import FSDPParallelDims
+from areal.utils.fsdp.parallelize import parallelize_model
 from areal.utils.save_load import get_state_dict_from_repo_id_or_path
 from areal.utils.ulysses import (
     set_ulysses_sequence_parallel_group,
@@ -60,12 +47,10 @@ class FSDPEngine(BaseHFEngine):
     def __init__(self, config: TrainEngineConfig):
         super().__init__(config)
         # FSDP options
-        self.fsdp_tp_device_mesh: DeviceMesh
         self.cpu_offload: CPUOffloadPolicy | None = None
 
-        self.dp_world_size: int
-        self.sp_world_size: int
-        self.tp_world_size: int
+        self.parallel_dims: FSDPParallelDims
+        self.world_mesh: DeviceMesh
 
         self.dp_group: dist.ProcessGroup
         self.sp_group: dist.ProcessGroup
@@ -84,7 +69,7 @@ class FSDPEngine(BaseHFEngine):
 
     @property
     def data_parallel_world_size(self) -> int:
-        return self.dp_world_size
+        return self.parallel_dims.dp
 
     def current_data_parallel_head(self) -> int:
         return self.dp_head
@@ -111,159 +96,35 @@ class FSDPEngine(BaseHFEngine):
 
         self.logger = logging.getLogger(f"[FSDP Engine Rank {dist.get_rank()}]")
 
-        self.parallel_strategy = self._make_parallel_strategy(parallel_strategy)
+        parallel_strategy = self._make_parallel_strategy(parallel_strategy)
 
-        self.dp_world_size = self.parallel_strategy.data_parallel_size
-        # This is Ulysses sequence parallelism
-        self.sp_world_size = self.parallel_strategy.context_parallel_size
-        self.tp_world_size = self.parallel_strategy.tensor_parallel_size
+        self.parallel_dims = FSDPParallelDims(
+            dp=parallel_strategy.dp_size,
+            sp=parallel_strategy.cp_size,
+            tp=parallel_strategy.tp_size,
+            ep=parallel_strategy.ep_size,
+            etp=parallel_strategy.etp_size,
+            world_size=self.world_size,
+        )
 
         self.logger.info(
-            f"Initializing device mesh with mode d{self.dp_world_size}s{self.sp_world_size}t{self.tp_world_size}."
+            f"Initializing device mesh with parallel dims {str(self.parallel_dims)}."
         )
 
-        self.fsdp_tp_device_mesh = init_device_mesh(
-            current_platform.device_type,
-            mesh_shape=(self.dp_world_size * self.sp_world_size, self.tp_world_size),
-            mesh_dim_names=("fsdp", "tp"),
-        )
+        self.world_mesh = self.parallel_dims.world_mesh
 
-        nd_device_mesh = init_device_mesh(
-            current_platform.device_type,
-            mesh_shape=(self.dp_world_size, self.sp_world_size, self.tp_world_size),
-            mesh_dim_names=("dp", "sp", "tp"),
-        )
+        self.dp_group = self.world_mesh["dp"].get_group()
+        self.sp_group = self.world_mesh["sp"].get_group()
 
-        self.dp_group = nd_device_mesh["dp"].get_group()
-        self.sp_group = nd_device_mesh["sp"].get_group()
-
-        # Combined model parallel group (tp + sp)
-        nd_device_mesh["sp", "tp"]._flatten(mesh_dim_name="mp")
-        self.mp_group = nd_device_mesh["mp"].get_group()
+        # Sequence and model parallel group (sp+tp)
+        self.mp_group = self.world_mesh["sp_tp"].get_group()
 
         self.rank = dist.get_rank()
 
-        self.dp_head = int(nd_device_mesh["mp"].mesh[0].item())
+        self.dp_head = int(self.world_mesh["sp_tp"].mesh[0].item())
         self.dp_rank = dist.get_rank(self.dp_group)
 
         self.logger.info(f"Data parallel head {self.dp_head} and rank {self.dp_rank}")
-
-    def apply_tensor_parallel(self, device_mesh: DeviceMesh):
-        num_attention_heads: int
-        num_key_value_heads: int
-        try:
-            num_attention_heads, num_key_value_heads = (
-                self.model.config.num_attention_heads,  # type: ignore
-                self.model.config.num_key_value_heads,  # type: ignore
-            )
-        except AttributeError:
-            num_attention_heads, num_key_value_heads = (
-                self.model.config.text_config.num_attention_heads,  # type: ignore
-                self.model.config.text_config.num_key_value_heads,  # type: ignore
-            )
-
-        if (
-            num_attention_heads % self.tp_world_size != 0
-            or num_key_value_heads % self.tp_world_size != 0
-        ):
-            raise ValueError(
-                f"num_attention_heads {num_attention_heads} and num_key_value_heads {num_key_value_heads} must be divisible by tensor_parallel_size {self.tp_world_size}"
-            )
-
-        if not isinstance(self.model.model, nn.Module):
-            raise RuntimeError("Model does not have the required submodule 'model'.")
-
-        # For model or model.language_model
-        model_tp_plan: dict[str, ParallelStyle] = {
-            "embed_tokens": RowwiseParallel(
-                input_layouts=Replicate(),
-                output_layouts=Shard(1),
-                use_local_output=False,
-            ),
-            "layers.*.input_layernorm": SequenceParallel(),
-            # All-gather
-            "layers.*.self_attn": PrepareModuleInput(
-                input_kwarg_layouts={"hidden_states": Shard(1)},
-                desired_input_kwarg_layouts={"hidden_states": Replicate()},
-            ),
-            "layers.*.self_attn.q_proj": ColwiseParallel(),
-            "layers.*.self_attn.k_proj": ColwiseParallel(),
-            "layers.*.self_attn.v_proj": ColwiseParallel(),
-            # special q/k norm for qwen3
-            "layers.*.self_attn.q_norm": NoParallel(),
-            "layers.*.self_attn.k_norm": NoParallel(),
-            # Reduce in RowwiseParallel, Scatter by Shard(1)
-            "layers.*.self_attn.o_proj": RowwiseParallel(
-                output_layouts=Shard(1),
-                use_local_output=False,
-            ),
-            "layers.*.post_attention_layernorm": SequenceParallel(),
-            # All-gather
-            "layers.*.mlp": PrepareModuleInput(
-                input_layouts=Shard(1),
-                desired_input_layouts=Replicate(),
-            ),
-            "layers.*.mlp.gate_proj": ColwiseParallel(),
-            "layers.*.mlp.up_proj": ColwiseParallel(),
-            # Reduce in RowwiseParallel, Scatter by Shard(1)
-            "layers.*.mlp.down_proj": RowwiseParallel(
-                output_layouts=Shard(1),
-                use_local_output=False,
-            ),
-            "norm": SequenceParallel(),
-        }
-
-        if is_gemma3_model(self.model_config.model_type):
-            model_tp_plan["layers.*.pre_feedforward_layernorm"] = SequenceParallel()
-            model_tp_plan["layers.*.post_feedforward_layernorm"] = SequenceParallel()
-
-        # For root module
-        root_tp_plan: dict[str, ParallelStyle] = {
-            # All-gather
-            "lm_head": ColwiseParallel(
-                input_layouts=Shard(1),
-                output_layouts=Replicate(),
-            ),
-        }
-
-        if self.is_vision_model:
-            if self.model_config.model_type not in VALID_VISION_MODELS:
-                self.logger.warning(
-                    f"Vision model type {self.model_config.model_type} not in supported list {VALID_VISION_MODELS}."
-                )
-
-            if isinstance(self.model.model.language_model, nn.Module):
-                # For vision-language models, avoid sharding the embedding layer because
-                # the visual components access it without tensor parallelism support.
-                # Instead, configure the first transformer layer to handle input
-                # sharding properly.
-                model_tp_plan.pop("embed_tokens", None)
-                model_tp_plan["layers.0"] = PrepareModuleInput(
-                    input_layouts=Replicate(),
-                    desired_input_layouts=Shard(1),
-                )
-
-                parallelize_module(
-                    self.model.model.language_model,
-                    device_mesh=device_mesh,
-                    parallelize_plan=model_tp_plan,
-                )
-            else:
-                self.logger.warning(
-                    f"Vision model does not have the required submodule 'model.language_model'."
-                )
-        else:
-            parallelize_module(
-                self.model.model,
-                device_mesh=device_mesh,
-                parallelize_plan=model_tp_plan,
-            )
-
-        self.model = parallelize_module(
-            self.model,
-            device_mesh=device_mesh,
-            parallelize_plan=root_tp_plan,
-        )
 
     def initialize(self, addr: str | None, ft_spec: FinetuneSpec | None):
         # Initialize distributed enviroments and load model.
@@ -279,31 +140,27 @@ class FSDPEngine(BaseHFEngine):
         # Monkey patch: replace attention's forward() with Ulysses variant.
         apply_monkey_patch(
             model=self.model,
-            ulysses_sp_size=self.sp_world_size,
+            ulysses_sp_size=self.parallel_dims.sp,
         )
-
-        if self.tp_world_size > 1:
-            self.apply_tensor_parallel(self.fsdp_tp_device_mesh["tp"])
 
         # sharding_strategy = ShardingStrategy.FULL_SHARD
         # Simple auto wrap policy
-        mixed_precision_policy = MixedPrecisionPolicy(
-            param_dtype=getattr(torch, self.config.dtype),
-            reduce_dtype=getattr(torch, self.config.grad_reduce_dtype),
-            cast_forward_inputs=True,
-        )
         self.cpu_offload = (
             CPUOffloadPolicy() if self.config.fsdp.offload_params else None
         )
-        fsdp_kwargs = {
-            "mesh": self.fsdp_tp_device_mesh["fsdp"],
-            "mp_policy": mixed_precision_policy,
-            "offload_policy": self.cpu_offload,
-            "reshard_after_forward": True,
-        }
         tik = time.perf_counter()
-        apply_fsdp2(self.model, fsdp_kwargs, self.config.fsdp.wrap_policy)
-        self.logger.info(f"Applying FSDP2 time: {time.perf_counter() - tik}")
+        parallelize_model(
+            self.model,
+            config=self.config,
+            model_config=self.model_config,
+            nd_device_mesh=self.world_mesh,
+            parallel_dims=self.parallel_dims,
+            cpu_offload=self.cpu_offload,
+            wrap_policy=self.config.fsdp.wrap_policy,
+        )
+        self.logger.info(
+            f"Applying FSDP2 with N-D parallelism for {time.perf_counter() - tik:.2f} seconds"
+        )
 
         self.create_optimizer(ft_spec)
         self.initialized = True
@@ -476,7 +333,7 @@ class FSDPEngine(BaseHFEngine):
         assert self.optimizer_config is not None
         assert self.lr_scheduler is not None
 
-        if self.sp_world_size > 1:
+        if self.parallel_dims.sp > 1:
             set_ulysses_sequence_parallel_group(self.sp_group)
 
         self.optimizer.zero_grad()
@@ -499,7 +356,7 @@ class FSDPEngine(BaseHFEngine):
             mb_list.padding_lengths, mb_list.padded_mbs, mb_list.mbs
         ):
             ulysses_pad_size = 0
-            if self.sp_world_size > 1:
+            if self.parallel_dims.sp > 1:
                 input_ids = padded_mb_input["input_ids"]
                 position_ids = padded_mb_input.get("position_ids", None)
 
@@ -508,7 +365,9 @@ class FSDPEngine(BaseHFEngine):
                         ulysses_input_ids,
                         ulysses_position_ids,
                         ulysses_pad_size,
-                    ) = ulysses_pad(input_ids, position_ids, sp_size=self.sp_world_size)
+                    ) = ulysses_pad(
+                        input_ids, position_ids, sp_size=self.parallel_dims.sp
+                    )
                 else:
                     # Pad and slice the inputs
                     (
@@ -518,7 +377,7 @@ class FSDPEngine(BaseHFEngine):
                     ) = ulysses_pad_and_slice_inputs(
                         input_ids,
                         position_ids,
-                        sp_size=self.sp_world_size,
+                        sp_size=self.parallel_dims.sp,
                     )
 
                 if (
@@ -537,7 +396,7 @@ class FSDPEngine(BaseHFEngine):
             outputs = self.model(**inputs)
 
             logits = outputs.logits.squeeze(0)
-            if self.sp_world_size > 1:
+            if self.parallel_dims.sp > 1:
                 # Gather and remove Ulysses padding
                 gathered_logits = dist_F.all_gather(logits, group=self.sp_group)
                 logits = torch.cat(gathered_logits, dim=0)
@@ -549,15 +408,14 @@ class FSDPEngine(BaseHFEngine):
 
             # Scale loss for accumulation
             # To reverse the gradient averaging for SP groups
-            loss_scale *= self.dp_world_size
+            loss_scale *= self.parallel_dims.dp
 
             loss *= loss_scale
             loss.backward()
 
-        # NOTE: grad norm clip function is different
         grad_norm = fsdp2_clip_grad_norm(
             list(self.model.parameters()),
-            self.fsdp_tp_device_mesh,
+            self.world_mesh,
             max_norm=self.optimizer_config.gradient_clipping,
         )
 
@@ -583,7 +441,7 @@ class FSDPEngine(BaseHFEngine):
         loss_weight_fn: Callable[[Dict[str, Any]], torch.Tensor],
     ) -> torch.Tensor | None:
         """Evaluate on a batch."""
-        if self.sp_world_size > 1:
+        if self.parallel_dims.sp > 1:
             set_ulysses_sequence_parallel_group(self.sp_group)
 
         mb_list = self.prepare_mb_list(input_)
@@ -605,7 +463,7 @@ class FSDPEngine(BaseHFEngine):
             mb_list.padding_lengths, mb_list.padded_mbs, mb_list.mbs
         ):
             ulysses_pad_size = 0
-            if self.sp_world_size > 1:
+            if self.parallel_dims.sp > 1:
                 input_ids = padded_mb_input["input_ids"]
                 position_ids = padded_mb_input.get("position_ids", None)
 
@@ -614,7 +472,9 @@ class FSDPEngine(BaseHFEngine):
                         ulysses_input_ids,
                         ulysses_position_ids,
                         ulysses_pad_size,
-                    ) = ulysses_pad(input_ids, position_ids, sp_size=self.sp_world_size)
+                    ) = ulysses_pad(
+                        input_ids, position_ids, sp_size=self.parallel_dims.sp
+                    )
                 else:
                     # Pad and slice the inputs
                     (
@@ -624,7 +484,7 @@ class FSDPEngine(BaseHFEngine):
                     ) = ulysses_pad_and_slice_inputs(
                         input_ids,
                         position_ids,
-                        sp_size=self.sp_world_size,
+                        sp_size=self.parallel_dims.sp,
                     )
 
                 if (
@@ -643,7 +503,7 @@ class FSDPEngine(BaseHFEngine):
             outputs = self.model(**inputs)
 
             logits = outputs.logits.squeeze(0)
-            if self.sp_world_size > 1:
+            if self.parallel_dims.sp > 1:
                 # Gather and remove Ulysses padding
                 gathered_logits = dist_F.all_gather(logits, group=self.sp_group)
                 logits = torch.cat(gathered_logits, dim=0)
@@ -671,7 +531,7 @@ class FSDPEngine(BaseHFEngine):
         aggregate_fn: Callable[[List[Any]], Any] = torch.cat,
     ) -> Any | None:
         """Forward pass with optional post-processing."""
-        if self.sp_world_size > 1:
+        if self.parallel_dims.sp > 1:
             set_ulysses_sequence_parallel_group(self.sp_group)
 
         cu_seqlens = pack_tensor_dict(input_)["cu_seqlens"]
@@ -688,7 +548,7 @@ class FSDPEngine(BaseHFEngine):
             mb_list.padding_lengths, mb_list.padded_mbs, mb_list.mbs
         ):
             ulysses_pad_size = 0
-            if self.sp_world_size > 1:
+            if self.parallel_dims.sp > 1:
                 input_ids = padded_mb_input["input_ids"]
                 position_ids = padded_mb_input.get("position_ids", None)
 
@@ -697,7 +557,9 @@ class FSDPEngine(BaseHFEngine):
                         ulysses_input_ids,
                         ulysses_position_ids,
                         ulysses_pad_size,
-                    ) = ulysses_pad(input_ids, position_ids, sp_size=self.sp_world_size)
+                    ) = ulysses_pad(
+                        input_ids, position_ids, sp_size=self.parallel_dims.sp
+                    )
                 else:
                     # Pad and slice the inputs
                     (
@@ -707,7 +569,7 @@ class FSDPEngine(BaseHFEngine):
                     ) = ulysses_pad_and_slice_inputs(
                         input_ids,
                         position_ids,
-                        sp_size=self.sp_world_size,
+                        sp_size=self.parallel_dims.sp,
                     )
 
                 if (
@@ -726,7 +588,7 @@ class FSDPEngine(BaseHFEngine):
             outputs = self.model(**inputs)
 
             logits = outputs.logits.squeeze(0)
-            if self.sp_world_size > 1:
+            if self.parallel_dims.sp > 1:
                 # Gather and remove Ulysses padding
                 gathered_logits = dist_F.all_gather(logits, group=self.sp_group)
                 logits = torch.cat(gathered_logits, dim=0)
